refactor: lint functions
