+++ packages/f2elint/package.json
@@ -6,7 +6,7 @@
   "main": "./lib/index.js",
   "scripts": {
     "dev": "npm run copyfiles && tsc -w",
-    "build": "npm run copyfiles && tsc ",
+    "build": "rm -rf lib && npm run copyfiles && tsc",
     "copyfiles": "copyfiles -a -u 1 \"src/config/**\" lib",
     "link": "npm run build && npm link && npm link -g f2elint",
     "test": "npm run build && jest",
@@ -53,19 +53,19 @@
   "devDependencies": {
     "@types/command-exists": "^1.2.0",
     "@types/cross-spawn": "^6.0.2",
-    "@types/ejs": "^3.0.5",
-    "@types/eslint": "^7.2.6",
-    "@types/fs-extra": "^9.0.4",
-    "@types/glob": "^7.1.3",
-    "@types/lodash": "^4.14.165",
+    "@types/ejs": "^3.1.0",
+    "@types/eslint": "^8.4.1",
+    "@types/fs-extra": "^9.0.13",
+    "@types/glob": "^7.2.0",
+    "@types/lodash": "^4.14.180",
     "@types/node": "^14.14.11",
-    "@types/prettier": "^2.1.5",
-    "@types/stylelint": "^9.10.1",
-    "@types/text-table": "^0.2.1",
+    "@types/prettier": "^2.4.4",
+    "@types/stylelint": "^14.0.0",
+    "@types/text-table": "^0.2.2",
     "copyfiles": "^2.4.1",
-    "jest": "^27.0.0",
+    "jest": "^27.5.1",
     "nyc": "^15.1.0",
-    "typescript": "^4.1.2"
+    "typescript": "^4.6.2"
   },
   "dependencies": {
     "@babel/core": "^7.16.0",
@@ -98,7 +98,7 @@
     "inquirer": "^7.3.3",
     "is-docker": "^2.1.1",
     "lodash": "^4.17.20",
-    "markdownlint": "^0.22.0",
+    "markdownlint": "^0.25.1",
     "markdownlint-config-ali": "^0.1.1",
     "markdownlint-rule-helpers": "^0.13.0",
     "ora": "^5.1.0",
