+++ packages/f2elint/src/actions/scan.ts
@@ -2,10 +2,12 @@ import path from 'path';
 import fs from 'fs-extra';
 import glob from 'glob';
 import prettier from 'prettier';
-import * as eslint from '../lints/eslint';
-import { getLintConfig, formatResults } from '../lints/stylelint';
+import { getMarkdownlintConfig, formatMarkdownlintResults } from '../lints/markdownlint';
+import markdownLintRuleHelpers from 'markdownlint-rule-helpers';
+import { getESLintConfig, formatESLintResults } from '../lints/eslint';
+import { getStylelintConfig, formatStylelintResults } from '../lints/stylelint';
+import { ESLint } from 'eslint';
 import stylelint from 'stylelint';
-import * as markdownLint from '../lints/markdownlint';
 import {
   PKG_NAME,
   ESLINT_FILE_EXT,
@@ -15,6 +17,7 @@ import {
   PRETTIER_IGNORE_PATTERN,
 } from '../utils/constants';
 import type { ScanOptions, ScanResult, PKG, Config, ScanReport } from '../types';
+import markdownlint from 'markdownlint';
 
 export default async (options: ScanOptions): Promise<ScanReport> => {
   const { cwd, include, quiet, fix, outputReport, config: scanConfig } = options;
@@ -53,10 +56,10 @@ export default async (options: ScanOptions): Promise<ScanReport> => {
   if (config.enableESLint !== false) {
     try {
       const files = getLintFiles(ESLINT_FILE_EXT);
-      const cli = new eslint.ESLint(eslint.getLintConfig(options, pkg, config));
+      const cli = new ESLint(getESLintConfig(options, pkg, config));
       const reports = await cli.lintFiles(files);
-      fix && (await eslint.ESLint.outputFixes(reports));
-      results = results.concat(eslint.formatResults(reports, quiet));
+      fix && (await ESLint.outputFixes(reports));
+      results = results.concat(formatESLintResults(reports, quiet));
     } catch (e) {
       runErrors.push(e);
     }
@@ -67,10 +70,10 @@ export default async (options: ScanOptions): Promise<ScanReport> => {
     try {
       const files = getLintFiles(STYLELINT_FILE_EXT);
       const data = await stylelint.lint({
-        ...getLintConfig(options, pkg, config),
+        ...getStylelintConfig(options, pkg, config),
         files,
       });
-      results = results.concat(formatResults(data.results, quiet));
+      results = results.concat(formatStylelintResults(data.results, quiet));
     } catch (e) {
       runErrors.push(e);
     }
@@ -82,11 +85,28 @@ export default async (options: ScanOptions): Promise<ScanReport> => {
       const files = options.files
         ? getLintFiles(MARKDOWN_LINT_FILE_EXT)
         : glob.sync('**/*.md', { cwd, ignore: 'node_modules/**' });
-      const data = await markdownLint.lint({
-        ...markdownLint.getLintConfig(options, pkg, config),
+      const result = await markdownlint.promises.markdownlint({
+        ...getMarkdownlintConfig(options, pkg, config),
         files,
       });
-      results = results.concat(markdownLint.formatResults(data, quiet));
+      // 修复
+      if (options.fix) {
+        for (const file in result) {
+          if (!Object.prototype.hasOwnProperty.call(result, file)) continue;
+
+          const fixes = result[file].filter((error) => error.fixInfo);
+
+          if (fixes.length > 0) {
+            const originalText = fs.readFileSync(file, 'utf8');
+            const fixedText = markdownLintRuleHelpers.applyFixes(originalText, fixes);
+            if (originalText !== fixedText) {
+              fs.writeFileSync(file, fixedText, 'utf8');
+              result[file] = result[file].filter((error) => !error.fixInfo);
+            }
+          }
+        }
+      }
+      results = results.concat(formatMarkdownlintResults(result, quiet));
     } catch (e) {
       runErrors.push(e);
     }
