+++ packages/f2elint/src/lints/stylelint/getStylelintConfig.ts
@@ -1,0 +1,41 @@
+import path from 'path';
+import fs from 'fs-extra';
+import glob from 'glob';
+import { LinterOptions } from 'stylelint';
+import { STYLELINT_IGNORE_PATTERN } from '../../utils/constants';
+import type { ScanOptions, PKG, Config } from '../../types';
+
+/**
+ * 获取 Stylelint 配置
+ */
+export function getStylelintConfig(opts: ScanOptions, pkg: PKG, config: Config): LinterOptions {
+  const { cwd, fix } = opts;
+  if (config.enableStylelint === false) return {} as any;
+
+  const lintConfig: any = {
+    fix: Boolean(fix),
+    allowEmptyInput: true,
+  };
+
+  if (config.stylelintOptions) {
+    // 若用户传入了 stylelintOptions，则用用户的
+    Object.assign(lintConfig, config.stylelintOptions);
+  } else {
+    // 根据扫描目录下有无lintrc文件，若无则使用默认的 lint 配置
+    const lintConfigFiles = glob.sync('.stylelintrc?(.@(js|yaml|yml|json))', { cwd });
+    if (lintConfigFiles.length === 0 && !pkg.stylelint) {
+      lintConfig.configBasedir = path.resolve(__dirname, '../../node_modules/');
+      lintConfig.config = {
+        extends: 'stylelint-config-ali',
+      };
+    }
+
+    // 根据扫描目录下有无lintignore文件，若无则使用默认的 ignore 配置
+    const ignoreFilePath = path.resolve(cwd, '.stylelintignore');
+    if (!fs.existsSync(ignoreFilePath)) {
+      lintConfig.ignorePattern = STYLELINT_IGNORE_PATTERN;
+    }
+  }
+
+  return lintConfig;
+}
