+++ packages/f2elint/tsconfig.json
@@ -10,9 +10,7 @@
     "resolveJsonModule": true,
     "declaration": true,
     "downlevelIteration": true,
-    "types": [
-      "node"
-    ],
+    "types": ["node"]
   },
-  "include": ["src"],
+  "include": ["src"]
 }
